requirements.txt
discord.py
python-dotenv
requests
pymongo
dnspython
tzdata